# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the Heracle personal training & fitness platform.
  title: Heracle Fitness API
  version: "1.0"
paths:
...
  /user/caltest:
    get:
      description: Fetches the Google Calendar list for the authenticated user using
        their stored access token
      produces:
      - application/json
      responses:
        "200":
          description: Google Calendar list
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Token expired
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User or token not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Test Google Calendar
      tags:
      - User
  /user/onboarding-status:
    get:
      description: Returns the onboarding completion status for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Onboarding status
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get onboarding status
      tags:
      - User
  /user/profile:
    get:
      description: Returns the authenticated user's profile with full details
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User
  /user/targets:
    post:
      consumes:
      - application/json
      description: Saves or updates nutritional targets (calories, macros) for the
        authenticated user
      parameters:
      - description: Target data
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated targets
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Save nutritional targets
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"